Code cleanup for FunctionalTests (#36582)
Part of #35884

+++ test/EFCore.Specification.Tests/TestModels/JsonQuery/JsonQueryData.cs
@@ -756,7 +756,7 @@ public class JsonQueryData : ISetSource
             TestDoubleCollection = [-1.23456789, 1.23456789, 0.0],
             TestDecimalCollection = [-1234567890.01M],
             TestGuidCollection = [new Guid("12345678-1234-4321-7777-987654321000")],
-            TestInt16Collection = [short.MinValue, (short)0, short.MaxValue],
+            TestInt16Collection = [short.MinValue, 0, short.MaxValue],
             TestInt32Collection = [int.MinValue, 0, int.MaxValue],
             TestInt64Collection =
             [
@@ -885,7 +885,7 @@ public class JsonQueryData : ISetSource
             TestDoubleCollection = [-1.23456789, 1.23456789, 0.0],
             TestDecimalCollection = [-1234567890.01M],
             TestGuidCollection = [new Guid("12345678-1234-4321-7777-987654321000")],
-            TestInt16Collection = [short.MinValue, (short)0, short.MaxValue],
+            TestInt16Collection = [short.MinValue, 0, short.MaxValue],
             TestInt32Collection = [int.MinValue, 0, int.MaxValue],
             TestInt64Collection =
             [
@@ -903,7 +903,7 @@ public class JsonQueryData : ISetSource
             TestTimeSpanCollection = [new TimeSpan(0, 10, 9, 8, 7), new TimeSpan(0, -10, 9, 8, 7)],
             TestDateOnlyCollection = [new DateOnly(2234, 1, 23), new DateOnly(5321, 1, 21)],
             TestTimeOnlyCollection = [new TimeOnly(21, 42, 23), new TimeOnly(17, 17, 27)],
-            TestUnsignedInt16Collection = [ushort.MinValue, (ushort)0, ushort.MaxValue],
+            TestUnsignedInt16Collection = [ushort.MinValue, 0, ushort.MaxValue],
             TestUnsignedInt32Collection = [uint.MinValue, 0, uint.MaxValue],
             TestUnsignedInt64Collection =
             [
@@ -1009,7 +1009,7 @@ public class JsonQueryData : ISetSource
             TestDoubleCollection = [-1.23456789, 1.23456789, 0.0],
             TestDecimalCollection = [-1234567890.01M],
             TestGuidCollection = [new Guid("12345678-1234-4321-7777-987654321000")],
-            TestInt16Collection = [short.MinValue, (short)0, short.MaxValue],
+            TestInt16Collection = [short.MinValue, 0, short.MaxValue],
             TestInt32Collection = [int.MinValue, 0, int.MaxValue],
             TestInt64Collection =
             [
@@ -1027,7 +1027,7 @@ public class JsonQueryData : ISetSource
             TestTimeSpanCollection = [new TimeSpan(0, 10, 9, 8, 7), new TimeSpan(0, -10, 9, 8, 7)],
             TestDateOnlyCollection = [new DateOnly(3234, 1, 23), new DateOnly(4331, 1, 21)],
             TestTimeOnlyCollection = [new TimeOnly(13, 42, 23), new TimeOnly(7, 17, 25)],
-            TestUnsignedInt16Collection = [ushort.MinValue, (ushort)0, ushort.MaxValue],
+            TestUnsignedInt16Collection = [ushort.MinValue, 0, ushort.MaxValue],
             TestUnsignedInt32Collection = [uint.MinValue, 0, uint.MaxValue],
             TestUnsignedInt64Collection =
             [
@@ -1133,7 +1133,7 @@ public class JsonQueryData : ISetSource
             TestDoubleCollection = [-1.23456789, 1.23456789, 0.0],
             TestDecimalCollection = [-1234567890.01M],
             TestGuidCollection = [new Guid("12345678-1234-4321-7777-987654321000")],
-            TestInt16Collection = [short.MinValue, (short)0, short.MaxValue],
+            TestInt16Collection = [short.MinValue, 0, short.MaxValue],
             TestInt32Collection = [int.MinValue, 0, int.MaxValue],
             TestInt64Collection =
             [
@@ -1151,7 +1151,7 @@ public class JsonQueryData : ISetSource
             TestTimeSpanCollection = [new TimeSpan(0, 10, 9, 8, 7), new TimeSpan(0, -10, 9, 8, 7)],
             TestDateOnlyCollection = [new DateOnly(1638, 1, 23), new DateOnly(4321, 1, 21)],
             TestTimeOnlyCollection = [new TimeOnly(8, 22, 23), new TimeOnly(7, 27, 37)],
-            TestUnsignedInt16Collection = [ushort.MinValue, (ushort)0, ushort.MaxValue],
+            TestUnsignedInt16Collection = [ushort.MinValue, 0, ushort.MaxValue],
             TestUnsignedInt32Collection = [uint.MinValue, 0, uint.MaxValue],
             TestUnsignedInt64Collection =
             [
@@ -1237,7 +1237,7 @@ public class JsonQueryData : ISetSource
                 TestDoubleCollection = [-1.23456789, 1.23456789, 0.0],
                 TestDecimalCollection = [-1234567890.01M],
                 TestGuidCollection = new ReadOnlyCollection<Guid>([new Guid("12345678-1234-4321-7777-987654321000")]),
-                TestInt16Collection = [short.MinValue, (short)0, short.MaxValue],
+                TestInt16Collection = [short.MinValue, 0, short.MaxValue],
                 TestInt32Collection = [int.MinValue, 0, int.MaxValue],
                 TestInt64Collection = new ReadOnlyCollection<long>(
                 [
@@ -1348,7 +1348,7 @@ public class JsonQueryData : ISetSource
                 TestDoubleCollection = [-1.23456789, 1.23456789, 0.0],
                 TestDecimalCollection = [-1234567890.01M],
                 TestGuidCollection = new ReadOnlyCollection<Guid>([new Guid("12345678-1234-4321-7777-987654321000")]),
-                TestInt16Collection = [short.MinValue, (short)0, short.MaxValue],
+                TestInt16Collection = [short.MinValue, 0, short.MaxValue],
                 TestInt32Collection = [int.MinValue, 0, int.MaxValue],
                 TestInt64Collection = new ReadOnlyCollection<long>(
                 [
